@@ -30,14 +30,15 @@
   "dependencies": {
     "leaflet": "^1.9.4",
     "lodash.isequal": "^4.5.0",
-    "react": "^18.2.0",
-    "react-leaflet": "^4.2.1",
-    "react-native": "^0.74.2"
+    "react": "^19.0.0",
+    "react-leaflet": "^5.0.0",
+    "react-measure": "^2.5.2",
+    "react-native": "^0.78.0"
   },
   "peerDependencies": {
-    "expo-asset": "^10.x",
+    "expo-asset": "^11.x",
     "expo-file-system": "^17.x",
-    "react": "^18.x",
+    "react": "^18.2.0",
     "react-measure": "^2.5.x",
     "react-native": "*",
     "react-native-webview": "^13.x"
@@ -45,14 +46,14 @@
   "devDependencies": {
     "@types/leaflet": "^1.9.12",
     "@types/lodash.isequal": "^4.5.8",
-    "@types/node": "^20.14.8",
+    "@types/node": "^22.13.10",
     "@types/react": "^18.2.0",
-    "@types/react-dom": "^18.3.0",
+    "@types/react-dom": "^19.0.4",
     "@types/react-measure": "^2.0.12",
-    "expo": "^51.0.14",
-    "expo-asset": "~10.0.9",
-    "expo-file-system": "~17.0.1",
-    "react-native-webview": "13.8.6",
+    "expo": "^52.0.37",
+    "expo-asset": "~11.0.4",
+    "expo-file-system": "~18.0.11",
+    "react-native-webview": "13.13.4",
     "typescript": "^5.3.3"
   },
   "resolutions": {
